Move panels to TS
